==> src/01 data extraction.ipynb <==
# Data Extraction

arXiv makes the data available from an API (https://info.arxiv.org/help/api/index.html) where publications can be retrieved individually. For bulk access, a full dataset is uploaded to the Kaggle website and updated once a month (https://www.kaggle.com/datasets/Cornell-University/arxiv). This is the source that has been used for this project.

In this first notebook, the data is retrieved from the Kaggle API and stored locally<br>

The dataset retrieved has metadata about the arXiv publications and not the full content of each article.

import json
import pandas as pd
import pickle
import os

from kaggle.api.kaggle_api_extended import KaggleApi
import zipfile

api = KaggleApi()
api.authenticate()

!kaggle datasets download -d 'Cornell-University/arxiv'

with zipfile.ZipFile('arxiv.zip', 'r') as zip_ref:
    zip_ref.extractall('./')

!rm arxiv.zip

!mkdir ../data

!mv arxiv-metadata-oai-snapshot.json ../data

filepath = "../data/arxiv-metadata-oai-snapshot.json"

## example of data content
with open(filepath, "r") as f:
    line = f.readline()
json.loads(line)

items = []
with open(filepath, "r") as f:
    lines = f.readlines()
    print("Total number of items:", len(lines))
    for l in lines:
        item = json.loads(l)
        items.append(item)

raw_data = pd.DataFrame.from_records(items)

raw_data.head()

raw_data.info()

# check duplicates
raw_data = raw_data.merge(raw_data["id"].value_counts().rename("count_of_id"), how='left', left_on='id', right_index=True)
duplicates = raw_data[raw_data["count_of_id"]>1].sort_values("id")
duplicates["id"].unique()

# remove duplicates
raw_data.drop_duplicates("id", inplace=True)
raw_data.reset_index(drop=True, inplace=True)
raw_data.drop("count_of_id", axis=1, inplace=True)

raw_data.shape

The data contains articles from the 23rd of May 2007 until the 13th of January 2023

from_date = min(raw_data["update_date"])
to_date = max(raw_data["update_date"])
print(f'First date: {from_date}\nLast date: {to_date}')

Saving the file locally makes it easier to access later for future analysis.

raw_data.to_parquet("../data/raw.parquet.gzip")
