==> src/07 TFIDF NB LR train.ipynb <==
# TFIDF NB LR train

Naive Bayes and Logistic Regression models are trained using TF-IDF vectors as input. The result of each model is used to obtain preditions from the validation set and analyse the performance of each classifier. 

import warnings
warnings.filterwarnings("ignore")

import numpy as np
import os
import pandas as pd
import pickle
from sklearn.naive_bayes import MultinomialNB
from sklearn.linear_model import LogisticRegression
from sklearn import metrics

from trainevalutils import *

target_names = pickle.load(open("../data/class_labels_dic.pkl", 'rb'))

OUTPUT_NAME = 'results_TFIDF_NBLR-google.pkl'

OUTPUT_DIR = '../results/'
for dir_ in [OUTPUT_DIR]:
    if not os.path.exists(dir_):
        os.makedirs(dir_)

X_train_1_1 = pickle.load(open('../data/wip/TFIDF/TFIDF_X_train_1_1.pkl', 'rb'))
X_train_1_2 = pickle.load(open('../data/wip/TFIDF/TFIDF_X_train_1_2.pkl', 'rb'))                       
y_train = pickle.load(open('../data/wip/TFIDF/TFIDF_y_train.pkl', 'rb'))
 
X_val_1_1 = pickle.load(open('../data/wip/TFIDF/TFIDF_X_val_1_1.pkl', 'rb'))
X_val_1_2 = pickle.load(open('../data/wip/TFIDF/TFIDF_X_val_1_2.pkl', 'rb'))                          
y_val = pickle.load(open('../data/wip/TFIDF/TFIDF_y_val.pkl', 'rb'))

def run_classifier(clf, model_name):
    results = []
    model_versions = ['1','2']
    X_train_sets = [X_train_1_1, X_train_1_2]
    X_val_sets = [X_val_1_1, X_val_1_2]
    for mv, Xtr, Xvl in zip(model_versions , X_train_sets, X_val_sets): 
        clf.fit(Xtr, y_train)
        name = model_name+mv
        title = name+'_val'
        print("\n"+title+"\n")
        result_val = evaluate(title, clf, Xvl, y_val)
        results.append(result_val)
        print("=====================================================\n")
    return results

print("\nResults of Naive Bayes on the validation set\n")
naivebayes = MultinomialNB()
results_nb = run_classifier(naivebayes, 'NBtfifd')

print("\nResults of Logistic Regression on the validation set\n")
logreg = LogisticRegression()
results_lr = run_classifier(logreg, 'LRtfidf')

df_result = pd.concat([pd.DataFrame(results_nb),
                       pd.DataFrame(results_lr)])

df_result

df_result.to_pickle(OUTPUT_DIR+OUTPUT_NAME)
