==> src/10 GridSearchCV weight calculation.ipynb <==
## Experimental class balancing

Experiements were made to deal with the class imbalance issue (see more in the next Notebook). As part of this, cutom class weights were calculated.

In this notebook, optimal class weights are calculated by means of a grid-search cross validation technique, using a tuning set of 100,000 rows sampled from the training dataset (see previous notebook). With this technique, the training data (which is the tuning set in this situation) is divided into a number of folds and for each fold the same model is trained iteratively on different subsets of the data, while being tested on a smaller portion of it. Different sets of parameters are tested this way. In this case, the parameters were represented by different combinations of weights. A total of 405 models were trained in this process to find the most performant set of weights out of 81 combinations.

See the weights obtained is use in the next notebook.

import warnings
warnings.filterwarnings("ignore")

import numpy as np
import pandas as pd
from sklearn.linear_model import LogisticRegression
from sklearn import metrics
from sklearn.model_selection import train_test_split
import seaborn as sns
import matplotlib.pyplot as plt
import pickle
from sklearn.model_selection import GridSearchCV
import itertools
from tqdm import tqdm

FILE = "../data/w2v-vectors_goog_pren.parquet.gzip"
data = pd.read_parquet(FILE)

train_idx = pickle.load(open("../data/wip/train_idx.pkl", 'rb'))

training_data = data.loc[train_idx]

SEED = 3742
N_SAMPLE = 100_000

tuning_data = training_data.sample(N_SAMPLE,random_state=SEED)

del data
del training_data

features = 'google-news_w2v_mean_prenorm'
target = 'class'

tuning_data[target] = tuning_data[target].map({2:0,7:1,6:2,5:3,1:3,3:3,4:3,0:3})

X_train = np.stack(tuning_data[features])
y_train = tuning_data[target]

y_train.value_counts()

from datetime import datetime

from sklearn.utils.class_weight import compute_class_weight

classes = np.unique(y_train)  # list of unique class labels

class_weights = compute_class_weight('balanced', classes= np.unique(y_train), y=y_train)

def expand_weight_range(value, k=3, amplitude=.1):
    '''
      Creates additional values around an original given value.

      value: the starting weight value
      k: the total numbe of weights to obtain
      amplitude: expressed as a fraction, it is the threshold
             considered when creating the range of values.

      Example: 
      expand_weight_range(value=100, k=3, amplitude=0.20)

      result:
      array([110., 100.,  90.])

      '''
    a = value + value * amplitude
    b = value - value * amplitude
    res = np.linspace(a,b,k)
    return res



ranges = [expand_weight_range(x) for x in class_weights]

catesian_pr_weights = list(itertools.product(*ranges))

weight_ranges = [dict(zip(classes,x)) for x in catesian_pr_weights]

len(weight_ranges)

start = datetime.now()
print(f"start: {start}")


gsc = GridSearchCV(
    estimator = LogisticRegression(),
    param_grid={
        'class_weight' : [x for x in weight_ranges]
    },
    scoring = 'f1_macro',
    cv = 5,
    verbose=1
)



grid_result = gsc.fit(X_train, y_train)


print(f"Best params: {grid_result.best_params_}")


pickle.dump(grid_result, open("tmp/grid_result.pkl", 'wb'))

end = datetime.now()
print(f"end: {end}")

grid_result = pickle.load(open("tmp/grid_result.pkl", 'rb'))

print(f"initial weights: {list(class_weights)}")
print(f"after gridsearchcv: {list(grid_result.best_params_['class_weight'].values())}")
