==> src/11 class imbalance techniques.ipynb <==
## Experiments to tackle class imbalance

In each of these experiments, only the training set was treated with rebalancing or resampling techniques. The validation set was left with its classes having the same proportional representation as the original data. The reason behind this is that in this way the validation set would simulate data in the real world, outside of the experiment.

import warnings
warnings.filterwarnings("ignore")

import numpy as np
import os
import pandas as pd
import pickle
from sklearn.naive_bayes import MultinomialNB
from sklearn.linear_model import LogisticRegression
from sklearn.preprocessing import MinMaxScaler
from sklearn import metrics

from trainevalutils import generate_model_report
from trainevalutils import *

from imblearn.over_sampling import RandomOverSampler
from imblearn.under_sampling import RandomUnderSampler
from imblearn.pipeline import Pipeline
from imblearn.over_sampling import SMOTE

SEED = 3742

OUTPUT_NAME = 'results_rescaling.pkl'

OUTPUT_DIR = '../results/'
for dir_ in [OUTPUT_DIR]:
    if not os.path.exists(dir_):
        os.makedirs(dir_)

FILE = "../data/w2v-vectors_goog_pren.parquet.gzip"
data = pd.read_parquet(FILE)

### load train - val indexes
train_idx = pickle.load(open("../data/wip/train_idx.pkl", 'rb'))
val_idx = pickle.load(open("../data/wip/val_idx.pkl", 'rb'))

### split data by indexes
training_data = data.loc[train_idx]
validation_data = data.loc[val_idx]

del data

features = 'google-news_w2v_mean_prenorm'
target = 'target'

X_train_raw = training_data[features]
X_val_raw = validation_data[features]

y_train = training_data[target]
y_val = validation_data[target]

# Stack
X_train_stacked = np.stack(X_train_raw)
X_val_stacked = np.stack(X_val_raw)

# handling negative values
scaler = MinMaxScaler()
X_train = scaler.fit_transform(X_train_stacked)
X_val = scaler.transform(X_val_stacked)

def run_classifier(clf, model_name, X_tr, y_tr):
    results = []
    clf.fit(X_tr, y_tr)
    title = model_name+'_val'
    print("\n"+title+"\n")
    result_val = evaluate(title, clf, X_val, y_val)
    results.append(result_val)
    print("=====================================================\n")
    return results

### Model LRw2vgoogCWbal
Logistic Regression with default class weight (balanced), using mean W2V embeddings.

This is a cost-sensitive learning approach with default balanced weights, calculated for each class as count of samples / (count of classes * count of samples in the class)

print("\nResults of Logistic Regression on the validation set\n")
logreg_CWbal = LogisticRegression(class_weight='balanced')
results_lr_CWbal = run_classifier(logreg_CWbal, 'LRw2vgoogCWbal', X_train, y_train)

with open('tmp/res/LRw2vgoogCWbal.pkl','wb') as f:
    pickle.dump(results_lr_CWbal, f)

### Model LRw2vgoogCWgscv
Logistic Regression with custom weights calculated using grid search cross validation in the previous Jupyter Notebook.

print("\nResults of Logistic Regression on the validation set\n")
grid_result = pickle.load(open("tmp/grid_result.pkl", 'rb'))
logreg_CWgscv = LogisticRegression(**grid_result.best_params_)
results_lr_CWgscv = run_classifier(logreg_CWgscv, 'LRw2vgoogCWgscv', X_train, y_train)

with open('tmp/res/LRw2vgoogCWgscv.pkl','wb') as f:
    pickle.dump(results_lr_CWgscv, f)

### Model LRw2vgoogUnder
Logistic Regression with random undersampling.

In this technique, random instances are removed from the majority class to balance the dataset. However, this approach may lead to a lot of loss of information, especially if the dataset is already small. In this specific case, because the smallest class (other), only has 5% of the total documents, therefore the data is reduced to 5% of ifs original size if undersampling is applied.

randomUnderSampler = RandomUnderSampler()
X_undersampled, y_undersampled = randomUnderSampler.fit_resample(X_train, y_train)

y_undersampled.value_counts()

print("\nResults of Logistic Regression on the validation set\n")
logreg_under= LogisticRegression()
results_lr_under = run_classifier(logreg_under, 'LRw2vgoogUnder', X_undersampled, y_undersampled)

with open('tmp/res/LRw2vgoogUnder.pkl','wb') as f:
    pickle.dump(results_lr_under, f)

### Model LRw2vgoogOver
Logistic Regression with random random over sampling.

This technique involves duplicating instances from the minority class randomly until the same size of the target majority class is achieved. This approach may lead to overfitting.

randomOverSampler = RandomOverSampler()
X_oversampled, y_oversampled = randomOverSampler.fit_resample(X_train, y_train)

y_oversampled.value_counts()

print("\nResults of Logistic Regression on the validation set\n")
logreg_over= LogisticRegression()
results_lr_over = run_classifier(logreg_over, 'LRw2vgoogOver', X_oversampled, y_oversampled)

with open('tmp/res/LRw2vgoogOver.pkl','wb') as f:
    pickle.dump(results_lr_over, f)

### Model LRw2vgoogSmote
Logistic Regression with SMOTE.

SMOTE (Synthetic Minority Over-sampling Technique) is a technique that generates synthetic samples for the minority class by creating new instances that are similar to existing ones. It is a popular method that can improve the performance of classifiers, especially when the minority class has a small number of samples.

smote = SMOTE(random_state=SEED)
X_smote, y_smote = smote.fit_resample(X_train, y_train)

pickle.dump(X_smote, open("tmp/X_LRw2vgoogSmote.pkl", 'wb'))
pickle.dump(y_smote, open("tmp/y_LRw2vgoogSmote.pkl", 'wb'))

X_smote = pickle.load(open("tmp/X_LRw2vgoogSmote.pkl", 'rb'))

y_smote = pickle.load(open("tmp/y_LRw2vgoogSmote.pkl", 'rb'))

y_smote.value_counts()

print("\nResults of Logistic Regression on the validation set\n")
logreg_smote = LogisticRegression()
results_lr_smote = run_classifier(logreg_smote, 'LRw2vgoogSmote', X_smote, y_smote)

with open('tmp/res/LRw2vgoogSmote.pkl','wb') as f:
    pickle.dump(results_lr_smote, f)

saved_res = []
files = os.listdir('tmp/res')
for f in files:
    with open('tmp/res/'+f, 'rb') as f:
        saved_res.append(pd.DataFrame(pickle.load(f)))

df_result = pd.concat(saved_res)
df_result

df_result.to_pickle(OUTPUT_DIR+OUTPUT_NAME)
